github-actions: do tarball generation with minimal feature set
This would be a good mimic when the tarball is created in a minimalized, clean environment,
but then it would be used for building every modules.

This tries to be similar what Kira's "build packages" job does, when
the tarball is built on the `syslog-ng-tarball` docker containter,
then passed on to all kinds of platforms for building.

The main motivation is to find build issues, caused by disabled modules during tarball generation.
See: #3464 commit: 9356ba91ce73ae394bd2a60edb4d16226ab145a2
("python: fix missing files from EXTRA_DIST in case Python is disabled")

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>